add  Laravel Permission by Spatie

+++ app/Http/Controllers/TaskController.php
@@ -18,11 +18,15 @@ class TaskController extends Controller
  
     public function create()
     {
+        $this->authorize('manage tasks');
+        
         return view('tasks.create');
     }
  
     public function store(StoreTaskRequest $request)
     {
+        $this->authorize('manage tasks');
+        
         Task::create($request->validated());
  
         return redirect()->route('tasks.index');
@@ -30,11 +34,15 @@ class TaskController extends Controller
  
     public function edit(Task $task)
     {
+        $this->authorize('manage tasks');
+        
         return view('tasks.edit', compact('task'));
     }
  
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        $this->authorize('manage tasks');
+        
         $task->update($request->validated());
  
         return redirect()->route('tasks.index');
@@ -42,6 +50,8 @@ class TaskController extends Controller
  
     public function destroy(Task $task)
     {
+        $this->authorize('manage tasks');
+        
         $task->delete();
  
         return redirect()->route('tasks.index');
